requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
qrcode==7.4.2
Pillow>=10.3.0
aiofiles==23.2.1
pyzbar==0.1.9
opencv-python>=4.10.0.84
numpy>=1.26.0

